feat(IdeaContrller + IdeaPolicy):added gate on indiidual cards , so only authorized person can change,delete or visit individual card

+++ app/Http/Controllers/IdeaController.php
@@ -9,6 +9,7 @@ use App\IdeaStatus;
 use App\Models\Idea;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class IdeaController extends Controller
 {
@@ -56,6 +57,7 @@ class IdeaController extends Controller
      */
     public function show(Idea $idea)
     {
+        Gate::authorize('workWith',$idea);
         return view('ideas.show', [
             'idea' => $idea,
         ]);
@@ -66,7 +68,7 @@ class IdeaController extends Controller
      */
     public function edit(Idea $idea)
     {
-        //
+         Gate::authorize('workWith',$idea);
     }
 
     /**
@@ -74,7 +76,7 @@ class IdeaController extends Controller
      */
     public function update(UpdateIdeaRequest $request, Idea $idea)
     {
-        //
+        Gate::authorize('workWith',$idea);
     }
 
     /**
@@ -82,6 +84,7 @@ class IdeaController extends Controller
      */
     public function destroy(Idea $idea)
     {
+         Gate::authorize('workWith',$idea);
         $idea->delete();
 
         return to_route('ideas.index');
